# docs/api/gateway/client.openapi.yaml
...
openapi: 3.0.3
info:
    title: ClientGatewayService API
    description: |-
        ClientGatewayService describes the authenticated REST API surface that the
         web frontend talks to via api-client at /api/v2/*.

         Auth is handled by gateway middleware — the user_id is injected from the
         Firebase token, NEVER sent by the client. Request messages here reflect
         what the frontend actually sends (no user_id fields).
    version: 0.0.1
paths:
...
    /users/me/activities:
        get:
            tags:
                - ClientGatewayService
            description: ===================== Activities =====================
            operationId: ClientGatewayService_ListActivities
            parameters:
                - name: limit
                  in: query
                  schema:
                    type: integer
                    format: int32
                - name: pageToken
                  in: query
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ListActivitiesGatewayResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /users/me/activities/stats:
        get:
            tags:
                - ClientGatewayService
            operationId: ClientGatewayService_GetActivityStats
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/GetActivityStatsGatewayResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /users/me/activities/{id}:
        get:
            tags:
                - ClientGatewayService
            operationId: ClientGatewayService_GetActivity
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/StandardizedActivity'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
        delete:
            tags:
                - ClientGatewayService
            operationId: ClientGatewayService_DeleteActivity
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content: {}
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
...
    /users/me/activity-photos/upload-url:
        post:
            tags:
                - ClientGatewayService
            operationId: ClientGatewayService_GetActivityPhotoUploadUrl
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/GetActivityPhotoUploadUrlGatewayRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/GetActivityPhotoUploadUrlGatewayResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
...
    /users/me/exercise-library:
        get:
            tags:
                - ClientGatewayService
            description: ===================== Exercise Library =====================
            operationId: ClientGatewayService_GetExerciseLibrary
            parameters:
                - name: q
                  in: query
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/GetExerciseLibraryGatewayResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /users/me/export:
        post:
            tags:
                - ClientGatewayService
            description: ===================== Data Export =====================
            operationId: ClientGatewayService_ExportData
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ExportDataGatewayResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
...
    /users/me/parse-fit:
        post:
            tags:
                - ClientGatewayService
            description: ===================== FIT File Parse =====================
            operationId: ClientGatewayService_ParseFitFile
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/ParseFitFileGatewayRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/StandardizedActivity'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
...
    /users/me/showcase-management/preferences:
        get:
            tags:
                - ClientGatewayService
            description: ===================== Showcase Management =====================
            operationId: ClientGatewayService_GetShowcasePreferences
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ShowcaseProfile'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
        put:
            tags:
                - ClientGatewayService
            operationId: ClientGatewayService_UpdateShowcasePreferences
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/UpdateShowcasePreferencesGatewayRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ShowcaseProfile'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /users/me/showcase-management/profile:
        get:
            tags:
                - ClientGatewayService
            operationId: ClientGatewayService_GetShowcaseSettings
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/GetShowcaseSettingsGatewayResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
        put:
            tags:
                - ClientGatewayService
            operationId: ClientGatewayService_UpdateShowcaseSettings
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/UpdateShowcaseSettingsGatewayRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ShowcaseProfile'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /users/me/showcase-management/profile/entries/{showcaseId}:
        post:
            tags:
                - ClientGatewayService
            operationId: ClientGatewayService_AddShowcaseEntry
            parameters:
                - name: showcaseId
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content: {}
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
        delete:
            tags:
                - ClientGatewayService
            operationId: ClientGatewayService_RemoveShowcaseEntry
            parameters:
                - name: showcaseId
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content: {}
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /users/me/showcase-management/profile/picture:
        post:
            tags:
                - ClientGatewayService
            operationId: ClientGatewayService_GetShowcaseProfilePictureUploadUrl
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/GetPictureUploadUrlGatewayRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/GetPictureUploadUrlGatewayResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /users/me/showcase-management/profile/slug:
        put:
            tags:
                - ClientGatewayService
            operationId: ClientGatewayService_UpdateShowcaseSlug
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/UpdateShowcaseSlugGatewayRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/UpdateShowcaseSlugGatewayResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /users/me/showcase-management/profile/views:
        get:
            tags:
                - ClientGatewayService
            operationId: ClientGatewayService_GetShowcaseProfileViewStats
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ShowcaseViewStats'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /users/me/showcase-management/roundup-settings:
        put:
            tags:
                - ClientGatewayService
            operationId: ClientGatewayService_UpdateRoundupSettings
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/UpdateRoundupSettingsGatewayRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ShowcaseProfile'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /users/me/showcase-management/roundup/{periodKey}/recompute:
        post:
            tags:
                - ClientGatewayService
            operationId: ClientGatewayService_RecomputeRoundup
            parameters:
                - name: periodKey
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ShowcaseRoundup'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /users/me/showcase-management/views:
        get:
            tags:
                - ClientGatewayService
            operationId: ClientGatewayService_ListShowcaseViewStats
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ListShowcaseViewStatsGatewayResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
...
components:
    schemas:
...
        GetActivityPhotoUploadUrlGatewayRequest:
            type: object
            properties:
                activityId:
                    type: string
                filename:
                    type: string
                contentType:
                    type: string
...
        GetActivityStatsGatewayResponse:
            type: object
            properties:
                totalActivities:
                    type: integer
                    format: int32
                totalShowcases:
                    type: integer
                    format: int32
                lastActivityAt:
                    type: string
                uploadsThisMonth:
                    type: integer
                    format: int32
                activitiesThisMonth:
                    type: integer
                    format: int32
                activitiesThisWeek:
                    type: integer
                    format: int32
                currentStreakDays:
                    type: integer
                    format: int32
                longestStreakDays:
                    type: integer
                    format: int32
                uploadsThisWeek:
                    type: integer
                    format: int32
...
        ListActivitiesGatewayResponse:
            type: object
            properties:
                activities:
                    type: array
                    items:
                        $ref: '#/components/schemas/StandardizedActivity'
                nextPageToken:
                    type: string
...